Move random subcommands to their respective group
